ramdisk userland: the libraries and commands a boot ramdisk needs
Stage 5.5 needs a root filesystem image, and the image needs a userland:
sh, mount, umount, sysenv, minix-service, procfs, grep, fsck_mfs,
loadramdisk. None of them had ever been compiled for aarch64, and none of
them built - not because of the architecture, but because the tree is
frozen on 2018 and the compiler is GCC 15. Nine of the ten failures are
the same for every port; two are LP64.

Age, diagnosed by a compiler that did not exist when the code was written:

  libterminfo	putchar cast to a wider function type, and a calloc()
		whose count and size are transposed. The casts are real
		undefined behaviour, not a style question: the call goes
		through a type with an extra parameter. Replaced with two
		small thunks, one of which carries the caller's function
		pointer in the args slot that _ti_puts() already has.
  libedit	a pointer compared against '\0' - which is the NULL test
		written twice, and never the empty-string test it means;
		strncpy() where the length is known and memcpy() is what
		is wanted; a strncat() whose bound is computed from the
		source string, so it bounds nothing.
  zlib		seventeen deliberate fall-throughs in inflate's state
		machine, now marked. Vendor code; upstream zlib later
		added the same markers.
  bin/sh	histcmd() has more locals live across its setjmp() than
		the compiler used to notice - argc, argv, lflg, pat and
		efp are now volatile like the rest; one fall-through in
		the parser; and _RETURN_INT(x) as ((x), 0) warns at every
		call site that discards the value, which is nearly all of
		them, so the zero now comes from a call.
  libasyn	sigprocmask(SIG_SETMASK, &mask, &mask) passes one object
		to two restrict-qualified parameters. Split in two.

LP64:

  libvtreefs	fs_getdents() bounded the directory position by ULONG_MAX,
		which said "must fit in an int" only while unsigned long
		was 32 bits wide. It is fed to int arithmetic below, so
		the bound is now INT_MAX.
  procfs	a pid and a service slot are carried in a cbdata_t, which
		is a void *. Wider than an int here, so the round trip
		goes through intptr_t.
  fsck.mfs	s_max_size is an i32_t on-disk field being set to LONG_MAX.
		mfs/super.c already clamps the same field to INT32_MAX and
		says why.

Classification: [возраст] with two of [порт]

+++ minix/commands/fsck.mfs/fsck.c
@@ -594,7 +594,13 @@ void rw_super(int put)
   if (sb.s_log_zone_size < 0) fatal("zone size < block size");
   if (sb.s_max_size <= 0) {
 	printf("warning: invalid max file size %d\n", sb.s_max_size);
-  	sb.s_max_size = LONG_MAX;
+	/*
+	 * s_max_size is an i32_t: an on-disk field, not a host long.  The
+	 * bound was written as LONG_MAX back when the two agreed, which is
+	 * no longer true under LP64.  mfs/super.c clamps to INT32_MAX for
+	 * the same reason.
+	 */
+  	sb.s_max_size = INT32_MAX;
   }
 }
