+++ minix/fs/procfs/tree.c
@@ -185,7 +185,12 @@ construct_pid_dirs(void)
 		 * process could keep open a file or directory across the owner
 		 * change, it might be able to access information it shouldn't.
 		 */
-		if (pid != (pid_t)get_inode_cbdata(node) ||
+		/*
+		 * The callback datum is a void *, and what is stored in it
+		 * here is a pid.  Under LP64 the pointer is wider than the
+		 * pid, so the round trip has to go through intptr_t.
+		 */
+		if (pid != (pid_t)(intptr_t)get_inode_cbdata(node) ||
 		    !check_owner(node, i))
 			delete_inode(node);
 	}
@@ -215,7 +220,7 @@ construct_pid_dirs(void)
 		make_stat(&stat, i, NO_INDEX);
 
 		node = add_inode(root, name, i, &stat, nr_pid_entries,
-		    (cbdata_t)pid);
+		    (cbdata_t)(intptr_t)pid);
 
 		if (node == NULL)
 			out_of_inodes();
